260304_1

+++ ProjectManagement.Shared/Base/ProjectAppStorage/NumericInputBase.cs
@@ -6,8 +6,8 @@
         public string DisplayName { get; set; } = string.Empty;
         public int SortOrder { get; set; } = 0;
 
-        public double? MinInputValue { get; set; }
-        public double? MaxInputValue { get; set; }
+        public decimal? MinInputValue { get; set; }
+        public decimal? MaxInputValue { get; set; }
         public string SectionKey { get; set; } = string.Empty;
     }
 }
