+++ ProjectManagement.Shared/DTO/ProjectAppStorage/Questions/Groups/NumericInputDto.cs
@@ -6,8 +6,8 @@ public class NumericInputDto
     public int Id { get; set; }
     public string DisplayName { get; set; } = string.Empty;
 
-    public double? MinInputValue { get; set; }
-    public double? MaxInputValue { get; set; }
+    public decimal? MinInputValue { get; set; }
+    public decimal? MaxInputValue { get; set; }
     public string SectionKey { get; set; } = string.Empty;
     public int SortOrder { get; set; } = 0;
 }
