+++ ProjectManagement.Shared/DTO/ProjectAppStorage/ResourceDto.cs
@@ -2,15 +2,17 @@ using ProjectManagement.Shared.Base.AppTenant;
 using ProjectManagement.Shared.DTO.App.Dataloader;
 using ProjectManagement.Shared.Enums;
 using System.Text.Json.Serialization;
+
 namespace ProjectManagement.Shared.DTO.ProjectAppStorage;
-public  enum ResourceSource
+
+public enum ResourceSource
 {
     Base, FromCodition
 }
-public class ResourceDto: ResourceDLBase
+
+public class ResourceDto : ResourceDLBase
 {
     public CalcResCost CalcResCost { get; set; } = new();
-
 
     public int Id { get; set; }
     public int? FolderId { get; set; }
@@ -21,14 +23,20 @@ public class ResourceDto: ResourceDLBase
     public List<ResourcePropertyBindDto> Properties { get; set; } = [];
     public List<RoleDTO> CostRole { get; set; } = [];
     public List<RoleDTO> CapRole { get; set; } = [];
-    public double? CostStorageValue { get; set; }
-    public double? CostUserValue { get; set; }
+
+    // كانت double -> صارت decimal
+    public decimal? CostStorageValue { get; set; }
+    public decimal? CostUserValue { get; set; }
+
     public string NameUserValue { get; set; } = string.Empty;
     public int? StatusId { get; set; }
     public int? ResourceTypeId { get; set; }
     public int? ResourceSortId { get; set; }
     public int? AccountId { get; set; }
-    [JsonIgnore] public double TotalCost => (double)((Data.Cost * (decimal)Data.Quantity.GetValueOrDefault()) + Data.BaseCost.GetValueOrDefault());
+
+    [JsonIgnore]
+    public decimal TotalCost =>
+        (Data.Cost * Data.Quantity.GetValueOrDefault()) + Data.BaseCost.GetValueOrDefault();
 
     [JsonIgnore] public ResourceSource ResourceSource = ResourceSource.Base;
     [JsonIgnore] public bool IsAdded { get; set; }
